requirements.txt
pandas
flask
python-docx